Remove option setting from .onLoad()
The package supported two redundant approaches to default Twist options
for token and workspace directory (but not workspace id), which was
leading to confusing behavior when using the package. This commit
removes the setting of options() from environment variables that took
place in .onLoad(). Setting will now happen in the `twist_*` functions
defined in api.R.